[SRC] Moving nomad_log.py to src root level
