+++ src/airflow/providers/nomad/templates/airflow.cfg
@@ -739,7 +739,8 @@ fab_logging_level = DEBUG
 #
 # Variable: AIRFLOW__LOGGING__LOGGING_CONFIG_CLASS
 #
-# logging_config_class = airflow.providers.nomad.executors.nomad_log.NOMAD_LOG_CONFIG
+# logging_config_class = airflow.providers.nomad.nomad_log.NOMAD_LOG_CONFIG
+# task_log_merge_with_stderr= false
 logging_config_class =
