+++ tests/unit/nomad/config/unit_tests.cfg
@@ -28,7 +28,9 @@ task_log_reader = nomad_log_handler
 #
 # Variable: AIRFLOW__LOGGING__LOGGING_CONFIG_CLASS
 #
-logging_config_class = airflow.providers.nomad.executors.nomad_log.NOMAD_LOG_CONFIG
+logging_config_class = airflow.providers.nomad.nomad_log.NOMAD_LOG_CONFIG
+
+task_log_merge_with_stderr = false
 
 [nomad]
 # Generic Airflow executor parallelism (should be higher than 0)
